Add files via upload
For this exercise, I examined the NIDS-CRAM dataset. NIDS-CRAM is the National Income Dynamics Study - Coronavirus Rapid Mobile Survey. The survey investigates the socioeconomic impacts of the national lockdown associated with the State of Disaster declared in South Africa in March 2020, and the social and economic consequences of the global Coronavirus pandemic. From March 23 through to early May 2020, South Africa was placed under an economic and social lockdown. During this period only essential services workers could physically attend work, and any other work involving physical or social interaction was suspended; impacting employment, incomes, and livelihoods. One of the key lines of questioning in the survey explored the extent to which the amount of time and earnings changed between February (pre-lockdown) and April(during lockdown.)

The NIDS-CRAM dataset is publicly available from the University of Cape Town's Data First Open Portal, where it is available for download as a Stata file. The data set under examination here considers the responses from 7,073 interviews for 56 of the 140 questions that were asked. The survey and links to its data can be found here: http://www.nids.uct.ac.za/about/nids-cram/nids-cram.